+++ src/main/java/by/intereson/ebookservice/entities/PartOfTheOrder.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 import static jakarta.persistence.GenerationType.SEQUENCE;
 
 @Data
@@ -21,20 +23,27 @@ public class PartOfTheOrder {
     @GeneratedValue(strategy = SEQUENCE, generator = SEQ_NAME)
     @SequenceGenerator(name = SEQ_NAME, sequenceName = SEQ_NAME, allocationSize = 1)
     private Long id;
+
     @Column(name = "SUM_PRICE")
-    private Double sumPrice;
+    private BigDecimal sumPrice;
+
     @Column(name = "PRICE_BOOK")
-    private Double price;
-    @Column(name = "QUANTITY",nullable = false)
+    private BigDecimal price;
+
+    @Column(name = "QUANTITY", nullable = false)
     private Integer quantity;
-    @Column(name = "BOOK_NAME",nullable = false)
+
+    @Column(name = "BOOK_NAME", nullable = false)
     private String bookName;
+
     @ManyToOne
     @JoinColumn(name = "ORDER_ID")
     private Order order;
+
     @ManyToOne
     @JoinColumn(name = "BOOK_ID", nullable = false)
     private Book book;
+
     @ManyToOne
     @JoinColumn(name = "SHOPPING_CART_ID")
     private ShoppingCart shoppingCart;
